[changed] Maintain focus inside of tray
This adds a new prop `maintainFocus` which defaults to `true`.
This makes the tray "modal" in that it will not allow focus to escape
from within it until it is explicitly closed via close button, escape key, or
clicking in the overlay area.

Previous behavior can be achieved by setting the `maintainFocus` prop
to `false`.  This defaults to `true` so that it can achieve a greater level of
accessibility support which I felt was valuable to everyone so it is opt-out
rather than opt-in.