'require' proper module names for 'react' and 'react-dom'
I know this is a generated file but I could not find the original spot to change this. Can we change these 'require's so they pull in the right module? where would I do that? this will prevent us from needing the fallback in public/javascripts/ReactDOM.js in canvas-lms